# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nightscout Multi-Tenant API
  description: |
    The Nightscout Multi-Tenant API provides comprehensive access to diabetes management data including 
    continuous glucose monitoring (CGM) readings, treatments, device status, and user profiles. 
    This API supports both modern JWT-based authentication for multi-tenant deployments and 
    legacy API_SECRET authentication for backwards compatibility.
  version: 3.0.0
  contact:
    name: Nightscout Project
    url: https://github.com/nightscout/cgm-remote-monitor
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://{subdomain}.nightscout.com/api
    description: Multi-tenant production server
    variables:
      subdomain:
        default: demo
        description: Tenant subdomain
  - url: https://your-nightscout-instance.com/api
    description: Single-tenant deployment
  - url: http://localhost:1337/api
    description: Local development server
security:
  - bearerAuth: []
  - apiSecret: []
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                  minLength: 8
                name:
                  type: string
                  minLength: 2
                role:
                  type: string
                  enum: [viewer, caregiver, admin]
                  default: viewer
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /tenants/register:
    post:
      tags:
        - Tenants
      summary: Register new tenant
      description: Create a new tenant organization
      operationId: registerTenant
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subdomain
                - adminEmail
                - adminPassword
                - adminName
              properties:
                subdomain:
                  type: string
                  pattern: '^[a-z0-9-]+$'
                  minLength: 3
                  maxLength: 63
                adminEmail:
                  type: string
                  format: email
                adminPassword:
                  type: string
                  format: password
                  minLength: 8
                adminName:
                  type: string
                  minLength: 2
                hCaptchaResponse:
                  type: string
                  description: hCaptcha verification token
      responses:
        '201':
          description: Tenant registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tenant:
                    $ref: '#/components/schemas/Tenant'
                  loginUrl:
                    type: string
                    format: uri
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
    apiSecret:
      type: apiKey
      in: header
      name: api-secret
      description: SHA-1 hashed API secret
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [viewer, caregiver, admin]
        tenantId:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Tenant:
      type: object
      properties:
        id:
          type: string
        subdomain:
          type: string
        isActive:
          type: boolean
        settings:
          type: object
          properties:
            units:
              type: string
              enum: [mg/dl, mmol/l]
            timeFormat:
              type: integer
              enum: [12, 24]
            language:
              type: string
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        status:
          type: integer
        result:
          type: string
          example: error
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...